requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e git+https://github.com/UiL-OTS-labs/django-shared-core.git@v1.1.0-beta-1#egg=uil-django-core  # via -r requirements.in
asgiref==3.2.10           # via django
babel==2.8.0              # via -r requirements.in
backports-datetime-fromisoformat==1.0.0  # via -r requirements.in
blessings==1.7            # via curtsies
bpython==0.19             # via -r requirements.in
certifi==2020.6.20        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
cryptography==3.1         # via -r requirements.in
curtsies==0.3.4           # via bpython
django-braces==1.14.0     # via -r requirements.in
django-csp==3.7           # via -r requirements.in
django-debug-toolbar==3.1  # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django-simple-menu==1.2.1  # via -r requirements.in
django==3.0.10            # via -r requirements.in, django-braces, django-csp, django-debug-toolbar, django-simple-menu
greenlet==0.4.16          # via bpython
idna==2.10                # via requests
mysqlclient==2.0.1        # via -r requirements.in
pillow==7.2.0             # via -r requirements.in
pip-tools==5.3.1          # via -r requirements.in
pycparser==2.20           # via cffi
pygments==2.7.1           # via bpython
pyjwt==1.7.1              # via -r requirements.in
pytz==2020.1              # via babel, django
requests==2.24.0          # via bpython
six==1.15.0               # via bpython, cryptography, django-braces, pip-tools
sqlparse==0.3.1           # via django, django-debug-toolbar
urllib3==1.25.10          # via requests
wcwidth==0.2.5            # via curtsies

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
